code cleanup, x64 fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@11163 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/MirOTR/libgcrypt-1.4.6/random/rndw32.c
@@ -344,7 +344,7 @@ read_mbm_data (void (*add)(const void*, size_t, enum random_origins),
         {
           if (debug_me)
             log_debug ("rndw32#read_mbm_data: got %d bytes\n",
-                       (int)sizeof (SharedData));
+                       sizeof (SharedData));
           (*add) (mbmDataPtr, sizeof (SharedData), requester);
           UnmapViewOfFile (mbmDataPtr);
         }
